Split out analysis_servers
